moving the authotp to login function
Moving authotp  inside login function,  when we run the code without adding the creds to config.ini getting error as  " binascii.Error: Non-base32 digit found".  there are 2 ways to to add credentials one direct add it to config.ini or click login the form will show up will input all the cred details and write it to config.ini which is easy way.